+++ cppApplicationOld/src/utils/intersect.cpp
@@ -4,17 +4,17 @@
 
 Intersect::Intersect(const Ray& ray) : ray(ray) {}
 
-std::optional<std::tuple<double, Eigen::Vector3d>> Intersect::operator()(const Triangle& triangle) const {
+std::optional<std::tuple<float, Eigen::Vector3f>> Intersect::operator()(const Triangle& triangle) const {
     return triangle.intersects(ray);
 }
 
-std::optional<std::tuple<double, Eigen::Vector3d>> Intersect::operator()(const Sphere& sphere) const {
+std::optional<std::tuple<float, Eigen::Vector3f>> Intersect::operator()(const Sphere& sphere) const {
     return sphere.intersects(ray);
 }
 
-std::optional<std::tuple<double, Eigen::Vector3d>> Intersect::operator()(const Mesh& mesh) const {
-    std::optional<std::tuple<double, Eigen::Vector3d>> nearest_hit;
-    double min_t = std::numeric_limits<double>::infinity();
+std::optional<std::tuple<float, Eigen::Vector3f>> Intersect::operator()(const Mesh& mesh) const {
+    std::optional<std::tuple<float, Eigen::Vector3f>> nearest_hit;
+    float min_t = std::numeric_limits<float>::infinity();
     for (const auto& triangle : mesh.triangles) {
         auto hit = triangle.intersects(ray);
         if (hit.has_value() && std::get<0>(hit.value()) < min_t) {
